Update to BEAST 2.7
Modified master scripts and codebase
Added more test examples
