--- MASTER_scripts/BDSIR_hetero_anyN.xml
@@ -1,38 +1,0 @@
-<beast version='2.0' namespace='master:master.model:master.conditions:master.outputs:master.postprocessors'>
-  <run spec='InheritanceTrajectory'
-       samplePopulationSizes="true"
-       verbosity="1">
-    <model spec='Model' id='model'>
-      <population spec='Population' id='S' populationName='S'/>
-      <population spec='Population' id='I' populationName='I'/>
-      <population spec='Population' id='R' populationName='R'/>
-      <population spec="Population" id="I_sample" populationName="I_sample"/>
-      <reaction spec='Reaction' reactionName="Infection" rate="0.005">
-	S + I -> 2I
-      </reaction>
-      <reaction spec='Reaction' reactionName="Recovery" rate="0.2">
-	I -> R
-      </reaction>
-      <reaction spec="Reaction" reactionName="I_sample" rate="0.1">
-	I -> I_sample
-      </reaction>
-    </model>
-    <initialState spec='InitState'>
-      <populationSize spec='PopulationSize' population='@S' size='199'/>
-      <lineageSeed spec='Individual' population='@I'/>
-    </initialState>
-
-    <lineageEndCondition spec='LineageEndCondition'
-			 nLineages="0"/>
-
-    <postSimCondition spec="LeafCountPostSimCondition"
-		      nLeaves="50" exact="false" exceedCondition="true"/>
-
-    <inheritancePostProcessor spec='LineageFilter'
-			      reactionName="I_sample"/>
-
-    <output spec='NewickOutput' fileName='SIRTree_hetero_anyN.newick.tree'/>
-    <output spec='NexusOutput' fileName='SIRTree_hetero_anyN.nexus.tree'/>
-    <output spec='JsonOutput' fileName='SIRTree_hetero_anyN.json'/>
-  </run>
-</beast>
