--- MASTER_scripts/BD_hetero_anyN.xml
@@ -1,43 +1,0 @@
-<beast version='2.0' namespace='master:master.model:master.steppers:master.conditions:master.postprocessors:master.outputs'>
-    <run spec='InheritanceTrajectory' samplePopulationSizes="true"
-	 verbosity="3" simulationTime='0.08'>
-
-        <model spec='Model' id='model'>
-            <population spec='Population' populationName='I' id='I'/>
-            <population spec='Population' populationName='R' id='R'/>
-            <population spec='Population' populationName='I_sampled' id='I_sampled'/>
-
-            <reaction spec='Reaction' reactionName="Birth" rate="114.75"> <!-- rate="1"> -->
-                I -> I + I
-            </reaction>
-            <reaction spec='Reaction' reactionName="Death" rate="13.5"> <!-- rate="0.5"> -->
-                I -> R
-            </reaction>
-            <reaction spec='Reaction' reactionName="Sampling" rate="31.5">  <!-- rate="0.2"> -->
-                I -> I_sampled
-            </reaction>
-        </model>
-
-        <initialState spec='InitState'>
-	  <lineageSeed spec='Individual' population='@I'/>
-
-          <populationSize spec='PopulationSize' population='@R' size='0'/>
-          <populationSize spec='PopulationSize' population='@I_sampled' size='0'/>
-        </initialState>
-
-        <populationEndCondition spec='PopulationEndCondition'
-				population='@I' threshold="0" exceedCondition="false"
-				isRejection="true" />
-	<inheritancePostProcessor spec='LineageFilter'
-				  reactionName="Sampling"/>
-	<postSimCondition spec="LeafCountPostSimCondition"
-				  nLeaves="5" exact="false"
-				  exceedCondition="true"/>
-	<postSimCondition spec="LeafCountPostSimCondition"
-				  nLeaves="500" exact="false"
-				  exceedCondition="false"/>
-        <output spec='NewickOutput' fileName='BDTree_hetero_anyN_MASTER.newick.tree'/>
-	<output spec='NexusOutput' fileName='BDTree_hetero_anyN_MASTER.nexus.tree'/>
-	<output spec='JsonOutput' fileName='BDTree_hetero_anyN_MASTER.json'/>
-    </run>
-</beast>
