--- master/CC_hetero_sim.xml
@@ -1,136 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?><beast beautitemplate='Standard' beautistatus='' namespace="beast.core:beast.evolution.alignment:beast.evolution.tree.coalescent:beast.core.util:beast.evolution.nuc:beast.evolution.operators:beast.evolution.sitemodel:beast.evolution.substitutionmodel:beast.evolution.likelihood" required="" version="2.4">
-
-
-    <data
-id="dummy"
-name="alignment">
-<!-- SEQ_DATA -->
-<!--
-    <sequence id="seq_10_0.27" taxon="10_0.27"
-     totalcount="4" value="gca"/> -->
-                    </data>
-
-<map name="Uniform" >beast.math.distributions.Uniform</map>
-<map name="Exponential" >beast.math.distributions.Exponential</map>
-<map name="LogNormal" >beast.math.distributions.LogNormalDistributionModel</map>
-<map name="Normal" >beast.math.distributions.Normal</map>
-<map name="Beta" >beast.math.distributions.Beta</map>
-<map name="Gamma" >beast.math.distributions.Gamma</map>
-<map name="LaplaceDistribution" >beast.math.distributions.LaplaceDistribution</map>
-<map name="prior" >beast.math.distributions.Prior</map>
-<map name="InverseGamma" >beast.math.distributions.InverseGamma</map>
-<map name="OneOnX" >beast.math.distributions.OneOnX</map>
-
-
-<run id="mcmc" spec="MCMC" chainLength="200000" sampleFromPrior="true">
-    <state id="state" storeEvery="10000">
-        <tree id="Tree.t:dummy" name="stateNode">
-            <trait id="dateTrait.t:dummy"
-		   spec="beast.evolution.tree.TraitSet"
-		   traitname="date-backward">
-	      <!-- DATE_DATA -->
-               <!-- 1_7.92=7.92,
-		2_6.74=6.74,
-		3_0.92=0.92,
-		4_3.85=3.85,
-		5_8.22=8.22,
-		6_4.48=4.48,
-		7_6.23=6.23,
-		8_6.34=6.34,
-		9_9.23=9.23,
-		10_0.27=0.27
-		-->
-		<taxa id="TaxonSet.dummy" spec="TaxonSet">
-                    <alignment idref="dummy"/>
-                </taxa>
-            </trait>
-            <taxonset idref="TaxonSet.dummy"/>
-        </tree>
-        <parameter id="clockRate.c:dummy" name="stateNode">1.0</parameter>
-        <parameter id="popSize.t:dummy" name="stateNode">$(popSize)</parameter>
-    </state>
-
-    <init id="RandomTree.t:dummy" spec="beast.evolution.tree.RandomTree" estimate="false" initial="@Tree.t:dummy" taxa="@dummy">
-        <populationModel id="ConstantPopulation0.t:dummy" spec="ConstantPopulation">
-            <parameter id="randomPopSize.t:dummy" name="popSize">$(popSize)</parameter>
-        </populationModel>
-    </init>
-
-    <distribution id="posterior" spec="util.CompoundDistribution">
-        <distribution id="prior" spec="util.CompoundDistribution">
-            <distribution id="CoalescentConstant.t:dummy" spec="Coalescent">
-                <populationModel id="ConstantPopulation.t:dummy" spec="ConstantPopulation" popSize="@popSize.t:dummy"/>
-                <treeIntervals id="TreeIntervals.t:dummy" spec="TreeIntervals" tree="@Tree.t:dummy"/>
-            </distribution>
-            <prior id="ClockPrior.c:dummy" name="distribution" x="@clockRate.c:dummy">
-                <Uniform id="Uniform.7" name="distr" upper="Infinity"/>
-            </prior>
-            <prior id="PopSizePrior.t:dummy" name="distribution"
-		   x="@popSize.t:dummy">
-	      <Uniform id="Uniform.8" name="distr" upper="Infinity"/>
-            </prior>
-        </distribution>
-        <distribution id="likelihood" spec="util.CompoundDistribution" useThreads="true">
-            <distribution id="treeLikelihood.dummy" spec="ThreadedTreeLikelihood" data="@dummy" tree="@Tree.t:dummy">
-                <siteModel id="SiteModel.s:dummy" spec="SiteModel">
-                    <parameter id="mutationRate.s:dummy" estimate="false" name="mutationRate">1.0</parameter>
-                    <parameter id="gammaShape.s:dummy" estimate="false" name="shape">1.0</parameter>
-                    <parameter id="proportionInvariant.s:dummy" estimate="false" lower="0.0" name="proportionInvariant" upper="1.0">0.0</parameter>
-                    <substModel id="JC69.s:dummy" spec="JukesCantor"/>
-                </siteModel>
-                <branchRateModel id="StrictClock.c:dummy" spec="beast.evolution.branchratemodel.StrictClockModel" clock.rate="@clockRate.c:dummy"/>
-            </distribution>
-        </distribution>
-    </distribution>
-<!--
-    <operator id="StrictClockRateScaler.c:dummy" spec="ScaleOperator" parameter="@clockRate.c:dummy" scaleFactor="0.75" weight="3.0"/>
-
-    <operator id="strictClockUpDownOperator.c:dummy" spec="UpDownOperator" scaleFactor="0.75" weight="3.0">
-        <up idref="clockRate.c:dummy"/>
-        <down idref="Tree.t:dummy"/>
-    </operator>
--->
-    <operator id="CoalescentConstantTreeScaler.t:dummy" spec="ScaleOperator" scaleFactor="0.5" tree="@Tree.t:dummy" weight="3.0"/>
-
-    <operator id="CoalescentConstantTreeRootScaler.t:dummy" spec="ScaleOperator" rootOnly="true" scaleFactor="0.5" tree="@Tree.t:dummy" weight="3.0"/>
-
-    <operator id="CoalescentConstantUniformOperator.t:dummy" spec="Uniform" tree="@Tree.t:dummy" weight="30.0"/>
-
-    <operator id="CoalescentConstantSubtreeSlide.t:dummy" spec="SubtreeSlide" tree="@Tree.t:dummy" weight="15.0"/>
-
-    <operator id="CoalescentConstantNarrow.t:dummy" spec="Exchange" tree="@Tree.t:dummy" weight="15.0"/>
-
-    <operator id="CoalescentConstantWide.t:dummy" spec="Exchange" isNarrow="false" tree="@Tree.t:dummy" weight="3.0"/>
-
-    <operator id="CoalescentConstantWilsonBalding.t:dummy" spec="WilsonBalding" tree="@Tree.t:dummy" weight="3.0"/>
-<!--
-    <operator id="PopSizeScaler.t:dummy" spec="ScaleOperator" parameter="@popSize.t:dummy" scaleFactor="0.75" weight="3.0"/>
--->
-    <logger id="tracelog" fileName="CC_hetero_sim.log" logEvery="200000" model="@posterior" sanitiseHeaders="true" sort="smart">
-        <log idref="posterior"/>
-        <log idref="likelihood"/>
-        <log idref="prior"/>
-        <log idref="treeLikelihood.dummy"/>
-        <log id="TreeHeight.t:dummy" spec="beast.evolution.tree.TreeHeightLogger" tree="@Tree.t:dummy"/>
-        <log idref="clockRate.c:dummy"/>
-        <log idref="popSize.t:dummy"/>
-        <log idref="CoalescentConstant.t:dummy"/>
-    </logger>
-
-    <logger id="screenlog" logEvery="200000">
-        <log idref="posterior"/>
-        <log id="ESS.0" spec="util.ESS" arg="@posterior"/>
-	<log idref="TreeHeight.t:dummy"/>
-	<log idref="popSize.t:dummy"/>
-        <log idref="CoalescentConstant.t:dummy"/>
-        <log idref="likelihood"/>
-        <log idref="prior"/>
-    </logger>
-
-    <logger id="treelog.t:dummy" fileName="output.tree" logEvery="200000" mode="tree">
-        <log id="TreeWithMetaDataLogger.t:dummy" spec="beast.evolution.tree.TreeWithMetaDataLogger" tree="@Tree.t:dummy"/>
-    </logger>
-
-</run>
-
-</beast>
