+++ src/beast/base/evolution/branchratemodel/FixedClockModel.java
@@ -1,0 +1,7 @@
+package beast.base.evolution.branchratemodel;
+
+import beast.base.evolution.branchratemodel.StrictClockModel;
+
+public class FixedClockModel extends StrictClockModel {
+
+}
