+++ src/beast/base/evolution/likelihood/FixedTreeLikelihood.java
@@ -1,0 +1,10 @@
+package beast.base.evolution.likelihood;
+
+import beast.base.core.Description;
+import beast.base.evolution.likelihood.GenericTreeLikelihood;
+
+@Description("Dummy treeLikelihood for a fixed tree that provides a partition so it integrates "
+		+ "with BEAUti")
+public class FixedTreeLikelihood extends GenericTreeLikelihood {
+
+}
