--- src/beast/evolution/branchratemodel/FixedClockModel.java
@@ -1,5 +1,0 @@
-package beast.evolution.branchratemodel;
-
-public class FixedClockModel extends StrictClockModel {
-
-}
