--- src/beast/evolution/likelihood/FixedTreeLikelihood.java
@@ -1,9 +1,0 @@
-package beast.evolution.likelihood;
-
-import beast.core.Description;
-
-@Description("Dummy treeLikelihood for a fixed tree that provides a partition so it integrates "
-		+ "with BEAUti")
-public class FixedTreeLikelihood extends GenericTreeLikelihood {
-
-}
