+++ tree_model_adequacy.iml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="module" module-name="beast2" exported="" />
+    <orderEntry type="module" module-name="BeastFX" exported="" />
+    <orderEntry type="module" module-name="TreeStat2" exported="" />
+    <orderEntry type="module" module-name="bdsky-master" exported="" />
+    <orderEntry type="module" module-name="phylodynamics-master" exported="" />
+  </component>
+</module>
